back/requirements.txt
BeautifulSoup4==4.7.1
firebase-admin
google.cloud
nltk
treetaggerwrapper
wikipedia
Flask
